requirements.txt
flask==2.0.1
captcha==0.4
requests==2.26.0
pillow
pytesseract==0.3.8